fix: stale appState for MainMenu defaultItems rendered from Actions
