+++ src/components/mainMenu/DefaultItems.scss
